# Reduce Component blocking threshold memory usage by 2 bytes per component